+++ R/loo-helpers.R
@@ -184,6 +184,17 @@ loo_compare.brmsfit <- function(
 #'   \code{"marglik"} (log marginal likelihood).
 #' @param model_name Optional name of the model. If \code{NULL}
 #'   (the default) the name is taken from the call to \code{x}.
+#' @param overwrite Logical; Indicates if already stored fit 
+#'   indices should be overwritten. Defaults to \code{FALSE}.
+#' @param file Either \code{NULL} or a character string. In the latter case, the
+#'   fitted model object including the newly added criterion values is saved via
+#'   \code{\link{saveRDS}} in a file named after the string supplied in
+#'   \code{file}. The \code{.rds} extension is added automatically. Only applies
+#'   if new criteria were actually added via \code{add_ic} or if
+#'   \code{force_save} was set to \code{TRUE}.
+#' @param force_save Logical; only relevant if \code{file} is specified and
+#'   ignored otherwise. If \code{TRUE}, the fitted model object will be saved
+#'   regardless of whether new criteria were added via \code{add_ic}.
 #' @param ... Further arguments passed to the underlying 
 #'   functions computing the model fit criteria.
 #'   
@@ -207,7 +218,9 @@ add_criterion <- function(x, ...) {
 
 #' @rdname add_criterion
 #' @export
-add_criterion.brmsfit <- function(x, criterion, model_name = NULL, ...) {
+add_criterion.brmsfit <- function(x, criterion, model_name = NULL, 
+                                  overwrite = FALSE, file = NULL,
+                                  force_save = FALSE, ...) {
   unused_args <- intersect(names(list(...)), args_not_for_reloo())
   if (length(unused_args)) {
     unused_args <- collapse_comma(unused_args)
@@ -219,26 +232,37 @@ add_criterion.brmsfit <- function(x, criterion, model_name = NULL, ...) {
   } else {
     model_name <- deparse_combine(substitute(x)) 
   }
-  criterion <- unique(tolower(as.character(criterion)))
-  options <- c("loo", "waic", "kfold", "r2", "marglik")
+  criterion <- unique(as.character(criterion))
+  options <- c("loo", "waic", "kfold", "R2", "marglik")
   if (!length(criterion) || !all(criterion %in% options)) {
     stop2("Argument 'criterion' should be a subset of ",
           collapse_comma(options))
   }
-  # remove previously stored criterion objects
-  x[criterion] <- list(NULL)
+  if (!is.null(file)) {
+    file <- paste0(as_one_character(file), ".rds")
+  }
+  force_save <- as_one_logical(force_save)
+  overwrite <- as_one_logical(overwrite)
+  if (overwrite) {
+    # remove previously stored criterion objects
+    x[criterion] <- list(NULL)
+  }
+  new_criteria <- criterion[ulapply(x[criterion], is.null)]
   args <- list(x, ...)
   for (fun in intersect(criterion, c("loo", "waic", "kfold"))) {
     args$model_names <- model_name
     x[[fun]] <- do.call(fun, args)
   }
-  if ("r2" %in% criterion) {
+  if ("R2" %in% criterion) {
     args$summary <- FALSE
     x$R2 <- do.call(bayes_R2, args)
   }
   if ("marglik" %in% criterion) {
     x$marglik <- do.call(bridge_sampler, args)
   }
+  if (!is.null(file) && (force_save || length(new_criteria))) {
+    saveRDS(x, file = file)
+  } 
   x
 }