+++ R/priors.R
@@ -545,6 +545,7 @@ prior_effects.brmsterms <- function(x, data, ...) {
   def_scale_prior <- def_scale_prior(x, data)
   valid_dpars <- valid_dpars(x$family, bterms = x)
   prior <- empty_brmsprior()
+  cats <- if (is_ordinal(x)) extract_cat_names(x, data)
   for (dp in valid_dpars) {
     def_dprior <- def_dprior(x, dp, data = data)
     if (!is.null(x$dpars[[dp]])) {
@@ -552,9 +553,9 @@ prior_effects.brmsterms <- function(x, data, ...) {
       dp_prior <- prior_effects(
         x$dpars[[dp]], data = data,
         def_scale_prior = def_scale_prior,
-        def_dprior = def_dprior
+        def_dprior = def_dprior, cats = cats
       )
-    } else if (!is.null(x$fdpars[[dp]])){
+    } else if (!is.null(x$fdpars[[dp]])) {
       # parameter is fixed
       dp_prior <- empty_brmsprior()
     } else {
@@ -603,20 +604,18 @@ prior_effects.brmsterms <- function(x, data, ...) {
 }
 
 #' @export
-prior_effects.btl <- function(x, data, spec_intercept = TRUE,
-                              def_scale_prior = "", def_dprior = "", 
-                              ...) {
+prior_effects.btl <- function(x, data, ...) {
   # collect default priors for various kinds of effects
   # Args:
   #   spec_intercept: special parameter class for the Intercept?
   #   def_scale_prior: default prior for SD parameters
   # Return:
   #   An object of class brmsprior
-  prior_fe(x, data, def_dprior = def_dprior, spec_intercept = spec_intercept) +
-    prior_sp(x, data) +
-    prior_cs(x, data) +
-    prior_sm(x, data, def_scale_prior = def_scale_prior) + 
-    prior_gp(x, data, def_scale_prior = def_scale_prior)
+  prior_fe(x, data, ...) +
+    prior_sp(x, data, ...) +
+    prior_cs(x, data, ...) +
+    prior_sm(x, data, ...) + 
+    prior_gp(x, data, ...)
 }
 
 #' @export
@@ -624,7 +623,8 @@ prior_effects.btnl <- function(x, data, ...) {
   empty_brmsprior()
 }
 
-prior_fe <- function(bterms, data, spec_intercept = TRUE, def_dprior = "") {
+prior_fe <- function(bterms, data, spec_intercept = TRUE, 
+                     def_dprior = "", cats = NULL, ...) {
   # priors for population-level parameters
   # Args:
   #   spec_intercept: special parameter class for the Intercept? 
@@ -634,8 +634,20 @@ prior_fe <- function(bterms, data, spec_intercept = TRUE, def_dprior = "") {
   fixef <- colnames(data_fe(bterms, data)$X)
   px <- check_prefix(bterms)
   if (has_intercept(bterms$fe) && spec_intercept) {
-    prior <- prior + 
-      brmsprior(def_dprior, class = "Intercept", coef = "", ls = px)
+    coefs <- NULL
+    if (is_ordinal(bterms)) {
+      threshold <- bterms$family$threshold
+      if (isTRUE(threshold == "equidistant")) {
+        coefs <- "1"
+      } else {
+        stopifnot(!is.null(cats))
+        coefs <- cats[-length(cats)]
+      }
+    }
+    prior <- prior + brmsprior(
+      c(def_dprior, rep("", length(coefs))), 
+      class = "Intercept", coef = c("", coefs), ls = px
+    )
     fixef <- setdiff(fixef, "Intercept")
   }
   if (length(fixef)) {
@@ -645,7 +657,7 @@ prior_fe <- function(bterms, data, spec_intercept = TRUE, def_dprior = "") {
   prior
 }
 
-prior_sp <- function(bterms, data) {
+prior_sp <- function(bterms, data, ...) {
   # priors for special effects parameters
   # Returns:
   #   an object of class brmsprior
@@ -666,7 +678,7 @@ prior_sp <- function(bterms, data) {
   prior
 }
 
-prior_cs <- function(bterms, data) {
+prior_cs <- function(bterms, data, ...) {
   # priors for category spcific effects parameters
   # Returns:
   #   an object of class brmsprior
@@ -680,7 +692,7 @@ prior_cs <- function(bterms, data) {
   prior
 }
 
-prior_Xme <- function(meef, internal = FALSE) {
+prior_Xme <- function(meef, internal = FALSE, ...) {
   # default priors for hyper-parameters of noise-free variables
   # Returns:
   #   an object of class brmsprior
@@ -713,7 +725,7 @@ prior_Xme <- function(meef, internal = FALSE) {
   prior
 }
 
-prior_gp <- function(bterms, data, def_scale_prior) {
+prior_gp <- function(bterms, data, def_scale_prior, ...) {
   # default priors of gaussian processes
   # Returns:
   #   an object of class brmsprior
@@ -771,12 +783,12 @@ def_lscale_prior <- function(bterms, data, plb = 0.01, pub = 0.01) {
   out
 }
 
-prior_re <- function(ranef, def_scale_prior, internal = FALSE) {
+prior_re <- function(ranef, def_scale_prior, internal = FALSE, ...) {
   # priors for random effects parameters
   # Args:
   #   ranef: a list returned by tidy_ranef
   #   def_scale_prior: a character string defining the default
-  #                    prior for random effects SDs
+  #     prior for random effects SDs
   #   internal: see get_prior
   # Returns:
   #   an object of class brmsprior
@@ -834,15 +846,24 @@ prior_re <- function(ranef, def_scale_prior, internal = FALSE) {
   prior
 }
 
-prior_sm <- function(bterms, data, def_scale_prior) {
+prior_sm <- function(bterms, data, def_scale_prior, ...) {
   # priors for smooth terms
   # Args:
   #   def_scale_prior: a character string defining 
   #     the default prior for smooth SDs
   prior <- empty_brmsprior()
-  smterms <- all_terms(bterms[["sm"]])
-  if (length(smterms)) {
+  smef <- tidy_smef(bterms, data)
+  if (NROW(smef)) {
     px <- check_prefix(bterms)
+    # prior for the FE coefficients
+    Xs_names <- attr(smef, "Xs_names")
+    if (length(Xs_names)) {
+      prior <- prior + brmsprior(
+        class = "b", coef = c("", Xs_names), ls = px
+      )
+    }
+    # prior for SD parameters of the RE coefficients
+    smterms <- unique(smef$term)
     prior_strings <- c(def_scale_prior, rep("", length(smterms)))
     prior <- prior + brmsprior(
       class = "sds", coef = c("", smterms),