+++ R/stan-prior.R
@@ -21,7 +21,7 @@ stan_prior <- function(prior, class, coef = "", group = "",
   wsp <- wsp(nsp = wsp)
   prior_only <- identical(attr(prior, "sample_prior"), "only")
   prior <- subset2(prior, 
-                   class = class, coef = c(coef, ""), group = c(group, "")
+    class = class, coef = c(coef, ""), group = c(group, "")
   )
   if (class %in% c("sd", "cor")) {
     # only sd and cor parameters have global priors
@@ -91,7 +91,7 @@ stan_prior <- function(prior, class, coef = "", group = "",
   
   # generate stan prior statements
   class <- paste0(prefix, class, suffix)
-  if (any(with(prior, nchar(coef) & nchar(prior)))) {
+  if (any(with(prior, nzchar(coef) & nzchar(prior)))) {
     # generate a prior for each coefficient
     out <- sapply(
       seq_along(coef), individual_prior, 
@@ -368,7 +368,7 @@ stan_rngprior <- function(sample_prior, prior, par_declars,
   D$ub <- stan_extract_bounds(D$bounds, bound = "upper")
   Ibounds <- which(nzchar(D$bounds))
   if (length(Ibounds)) {
-    str_add(out$genC) <- " // use rejection sampling for truncated priors\n"
+    str_add(out$genC) <- "  // use rejection sampling for truncated priors\n"
     for (i in Ibounds) {
       wl <- if (nzchar(D$lb[i])) paste0(D$prior_par[i], " < ", D$lb[i])
       wu <- if (nzchar(D$ub[i])) paste0(D$prior_par[i], " > ", D$ub[i])